pyproject.toml
[project]
name = "complete-mcp-server-mastery-with-langchain-and-ollama"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.2",
    "matplotlib>=3.10.6",
    "mcp[cli]>=1.14.1",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scipy>=1.16.2",
]
